+++ src/frontend/apps/desk/src/features/header/index.ts
@@ -1,0 +1,1 @@
+export * from './Header';
