+++ src/frontend/apps/desk/src/features/index.ts
@@ -1,0 +1,4 @@
+export * from './header/';
+export * from './language/';
+export * from './menu/';
+export * from './teams/';
